feat: Headless middleware foe protecting internal routes
