+++ jsend/middleware.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"runtime"
 
+	"github.com/random-guys/go-siber/sessions"
 	"github.com/rs/zerolog"
 )
 
@@ -53,5 +54,20 @@ func Recoverer(env string) func(http.Handler) http.Handler {
 		}
 
 		return http.HandlerFunc(fn)
+	}
+}
+
+func Headless(store *sessions.Store) func(http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			type void struct{}
+			var empty void
+
+			// force a panic if you have to
+			LoadHeadless(store, r, empty)
+
+			// nothing to worry about
+			next.ServeHTTP(w, r)
+		})
 	}
 }
